+++ src/components/task-input.ts
@@ -1,0 +1,99 @@
+import { Validatable, validate } from "../util/validation.js";
+import { tasksState } from "../state/task-state.js";
+
+export class taskInput {
+  templateElement: HTMLTemplateElement;
+  hostElement: HTMLDivElement;
+  element: HTMLFormElement;
+  titleInputElement: HTMLInputElement;
+  descriptionInputElement: HTMLInputElement;
+  deadlineInputElement: HTMLInputElement;
+
+  constructor() {
+    this.templateElement = document.getElementById(
+      "task-input"
+    )! as HTMLTemplateElement;
+    this.hostElement = document.getElementById("app")! as HTMLDivElement;
+
+    const importedNode = document.importNode(
+      this.templateElement.content,
+      true
+    );
+    this.element = importedNode.firstElementChild as HTMLFormElement;
+    this.element.id = "user-input";
+
+    this.titleInputElement = this.element.querySelector(
+      "#title"
+    ) as HTMLInputElement;
+    this.descriptionInputElement = this.element.querySelector(
+      "#description"
+    ) as HTMLInputElement;
+    this.deadlineInputElement = this.element.querySelector(
+      "#deadline"
+    ) as HTMLInputElement;
+
+    this.configure();
+    this.attach();
+  }
+
+  private gatherUserInput(): [string, string, string] | void {
+    const enteredTitle = this.titleInputElement.value;
+    const enteredDescription = this.descriptionInputElement.value;
+    const enteredDeadline = this.deadlineInputElement.value;
+
+    const titleValidatable: Validatable = {
+      value: enteredTitle,
+      required: true,
+    };
+    const descriptionValidatable: Validatable = {
+      value: enteredDescription,
+      required: true,
+      minLength: 5,
+    };
+    const deadlineValidatable: Validatable = {
+      value: enteredDeadline,
+      required: true,
+      startDate: true,
+    };
+    if (
+      !validate(titleValidatable) ||
+      !validate(descriptionValidatable) ||
+      !validate(deadlineValidatable)
+    ) {
+      alert("入力値が不正です");
+      return;
+    } else {
+      return [enteredTitle, enteredDescription, enteredDeadline];
+    }
+  }
+
+  private clearInputs() {
+    this.titleInputElement.value = "";
+    this.descriptionInputElement.value = "";
+    this.deadlineInputElement.value = "";
+  }
+
+  private submitHandler(event: Event) {
+    event.preventDefault();
+    const userInput = this.gatherUserInput();
+    if (Array.isArray(userInput)) {
+      const [title, desc, deadline] = userInput;
+      console.log(title, desc, deadline);
+      tasksState.addTasks(
+        Math.random().toString(32).substring(2),
+        title,
+        desc,
+        deadline
+      );
+      this.clearInputs();
+    }
+  }
+
+  private configure() {
+    this.element.addEventListener("submit", this.submitHandler.bind(this));
+  }
+
+  private attach() {
+    this.hostElement.insertAdjacentElement("afterbegin", this.element);
+  }
+}
