+++ src/models/drag-and-drop.ts
@@ -1,0 +1,8 @@
+export interface Draggable {
+  dragStartHandler(event: DragEvent): void;
+}
+
+export interface DragTarget {
+  dragOverHandler(event: DragEvent): void;
+  dropHandler(event: DragEvent): void;
+}
