+++ src/models/task.ts
@@ -1,0 +1,9 @@
+export class Task {
+  constructor(
+    public id: string,
+    public title: string,
+    public description: string,
+    public deadline: string,
+    public status: "active" | "finished"
+  ) {}
+}
